move main hooks into main class
