+++ inc/functions.php
@@ -8,76 +8,6 @@
 namespace XMLSF;
 
 /**
- * Sitemap loaded
- *
- * Common actions to prepare sitemap loading.
- * Hooked into xmlsf_sitemap_loaded action.
- */
-function sitemap_loaded() {
-	// Prepare headers.
-	\add_filter( 'wp_headers', __NAMESPACE__ . '\headers' );
-
-	// Set the sitemap conditional flag.
-	xmlsf()->is_sitemap = true;
-
-	// Make sure we have the proper locale setting for calculations.
-	\setlocale( LC_NUMERIC, 'C' );
-
-	// Save a few db queries.
-	\add_filter( 'split_the_query', '__return_false' );
-
-	// Don't go redirecting anything from now on...
-	\remove_action( 'template_redirect', 'redirect_canonical' );
-
-	/** GENERAL MISC. PREPARATIONS */
-
-	// Prevent public errors breaking xml.
-	@\ini_set( 'display_errors', 0 ); // phpcs:ignore WordPress.PHP.NoSilencedErrors.Discouraged,WordPress.PHP.IniSet.display_errors_Disallowed
-
-	// Remove filters to prevent stuff like cdn urls for xml stylesheet and images.
-	\remove_all_filters( 'plugins_url' );
-	\remove_all_filters( 'wp_get_attachment_url' );
-	\remove_all_filters( 'image_downsize' );
-
-	// Remove actions that we do not need.
-	\remove_all_actions( 'widgets_init' );
-	\remove_all_actions( 'wp_footer' );
-}
-
-\add_action( 'xmlsf_sitemap_loaded', __NAMESPACE__ . '\sitemap_loaded' );
-\add_action( 'xmlsf_news_sitemap_loaded', __NAMESPACE__ . '\sitemap_loaded' );
-
-/**
- * Filter robots.txt rules
- *
- * @param string $output Default robots.txt content.
- *
- * @return string
- */
-function robots_txt( $output ) {
-
-	// CUSTOM ROBOTS.
-	$robots_custom = \get_option( 'xmlsf_robots' );
-	$output       .= $robots_custom ? $robots_custom . PHP_EOL : '';
-
-	// SITEMAPS.
-
-	$output .= PHP_EOL . '# XML Sitemap & Google News version ' . XMLSF_VERSION . ' - https://status301.net/wordpress-plugins/xml-sitemap-feed/' . PHP_EOL;
-	if ( 1 !== (int) \get_option( 'blog_public' ) ) {
-		$output .= '# XML Sitemaps are disabled because of this site\'s visibility settings.' . PHP_EOL;
-	} elseif ( ! namespace\sitemaps_enabled() ) {
-		$output .= '# No XML Sitemaps are enabled.' . PHP_EOL;
-	} else {
-		namespace\sitemaps_enabled( 'sitemap' ) && $output .= 'Sitemap: ' . namespace\sitemap_url() . PHP_EOL;
-		namespace\sitemaps_enabled( 'news' ) && $output    .= 'Sitemap: ' . namespace\sitemap_url( 'news' ) . PHP_EOL;
-	}
-
-	return $output;
-}
-
-\add_filter( 'robots_txt', __NAMESPACE__ . '\robots_txt' );
-
-/**
  * Get the public XML sitemap url.
  *
  * @since 5.4
@@ -162,7 +92,6 @@ function xml_stylesheet( $sitemap = false ) {
  * @return string|false
  */
 function get_stylesheet_url( $sitemap = false ) {
-
 	/**
 	 * GET STYLESHEET URL
 	 *
@@ -274,7 +203,6 @@ function headers( $headers ) {
  * @param string $feed            Feed type.
  */
 function load_template( $is_comment_feed, $feed ) {
-
 	/**
 	 * GET TEMPLATE FILE
 	 *
@@ -354,18 +282,6 @@ function sanitize_number( $number, $min = .1, $max = 1, $decimals = 1 ) {
 
 	return \number_format( $number, $decimals, '.', '' );
 }
-
-/**
- * Generator info
- */
-function generator() {
-	echo '<!-- generated-on="' . \esc_xml( \gmdate( 'c' ) ) . '" -->' . PHP_EOL;
-	echo '<!-- generator="XML Sitemap & Google News for WordPress" -->' . PHP_EOL;
-	echo '<!-- generator-url="https://status301.net/wordpress-plugins/xml-sitemap-feed/" -->' . PHP_EOL;
-	echo '<!-- generator-version="' . \esc_xml( XMLSF_VERSION ) . '" -->' . PHP_EOL;
-}
-
-add_action( 'xmlsf_generator', __NAMESPACE__ . '\generator' );
 
 /**
  * Clear cache metadata
